+++ Enderlook.EventManager/src/Containers.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.Runtime;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -26,30 +27,54 @@ namespace Enderlook.EventManager
         }
     }
 
-    internal readonly struct InvariantObjectAndGCHandle : IWeak
+    internal struct InvariantObjectAndGCHandle : IWeak
     {
+#if NET6_0
+        public DependentHandle Token;
+#else
         public readonly object Value;
         public readonly GCHandle Handle;
+#endif
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public InvariantObjectAndGCHandle(object value, object handle)
         {
+#if NET6_0
+            Token = new DependentHandle(handle, value);
+#else
             Value = value;
             Handle = GCHandle.Alloc(handle, GCHandleType.Weak);
+#endif
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Free() => Handle.Free();
+        public void Free()
+        {
+#if NET6_0
+            Token.Dispose();
+#else
+            Handle.Free();
+#endif
+        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool FreeIfIsCollected()
         {
+#if NET6_0
+            if (Token.Dependent is null)
+            {
+                Token.Dispose();
+                return true;
+            }
+            return false;
+            #else
             if (Handle.Target is null)
             {
                 Handle.Free();
                 return true;
             }
             return false;
+#endif
         }
     }
 
@@ -80,32 +105,56 @@ namespace Enderlook.EventManager
         }
     }
 
-    internal readonly struct InvariantObjectAndTAndGCHandle<T> : IWeak
+    internal struct InvariantObjectAndTAndGCHandle<T> : IWeak
     {
-        public readonly object Value;
         public readonly T ValueT;
+#if NET6_0
+        public DependentHandle Token;
+#else
+        public readonly object Value;
         public readonly GCHandle Handle;
+#endif
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public InvariantObjectAndTAndGCHandle(object value, object handle, T valueT)
         {
-            Value = value;
             ValueT = valueT;
+#if NET6_0
+            Token = new DependentHandle(handle, value);
+#else
+            Value = value;
             Handle = GCHandle.Alloc(handle, GCHandleType.Weak);
+#endif
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Free() => Handle.Free();
+        public void Free()
+        {
+#if NET6_0
+            Token.Dispose();
+#else
+            Handle.Free();
+#endif
+        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool FreeIfIsCollected()
         {
+#if NET6_0
+            if (Token.Dependent is null)
+            {
+                Token.Dispose();
+                return true;
+            }
+            return false;
+#else
             if (Handle.Target is null)
             {
                 Handle.Free();
                 return true;
             }
             return false;
+#endif
         }
     }